+++ rust/operator-binary/src/discovery.rs
@@ -1,8 +1,9 @@
-use crate::hbase_controller::{build_recommended_labels, hbase_version};
+use crate::hbase_controller::build_recommended_labels;
 use stackable_hbase_crd::{HbaseCluster, HbaseRole, HBASE_SITE_XML, HBASE_ZOOKEEPER_QUORUM};
 use stackable_operator::{
     builder::{ConfigMapBuilder, ObjectMetaBuilder},
-    error::{Error, OperatorResult},
+    commons::product_image_selection::ResolvedProductImage,
+    error::OperatorResult,
     k8s_openapi::api::core::v1::ConfigMap,
 };
 use std::collections::HashMap;
@@ -11,6 +12,7 @@ use std::collections::HashMap;
 pub fn build_discovery_configmap(
     hbase: &HbaseCluster,
     zookeeper_connect_string: &str,
+    resolved_product_image: &ResolvedProductImage,
 ) -> OperatorResult<ConfigMap> {
     let hbase_site_data: HashMap<String, Option<String>> = [(
         HBASE_ZOOKEEPER_QUORUM.to_string(),
@@ -25,7 +27,7 @@ pub fn build_discovery_configmap(
                 .ownerreference_from_resource(hbase, None, Some(true))?
                 .with_recommended_labels(build_recommended_labels(
                     hbase,
-                    hbase_version(hbase).map_err(|_| Error::MissingObjectKey { key: "version" })?,
+                    &resolved_product_image.app_version_label,
                     &HbaseRole::RegionServer.to_string(),
                     "discovery",
                 ))